+++ src/Attribute/SdcAsset.php
@@ -14,7 +14,7 @@ namespace Tito10047\UX\Sdc\Attribute;
 use Attribute;
 
 #[Attribute(Attribute::TARGET_CLASS | Attribute::IS_REPEATABLE)]
-final class Asset
+final class SdcAsset
 {
     /**
      * @param array<string, mixed> $attributes