+++ tests/Integration/AsSdcComponentIntegrationTest.php
@@ -15,7 +15,6 @@ use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\UX\TwigComponent\ComponentFactory;
 use Tito10047\UX\Sdc\Runtime\SdcMetadataRegistry;
 use Tito10047\UX\Sdc\Tests\Integration\Fixtures\Component\SdcComponent;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Tito10047\UX\Sdc\Tests\Integration\Fixtures\Component\SdcComponentWithAsset;
 
 class AsSdcComponentIntegrationTest extends KernelTestCase
@@ -35,7 +34,6 @@ class AsSdcComponentIntegrationTest extends KernelTestCase
         /** @var ComponentFactory $componentFactory */
         $componentFactory = $container->get('ux.twig_component.component_factory');
 
-        // This will fail unless we register SdcComponent in TestKernel
         $metadata = $componentFactory->metadataFor('SdcComponent');
 
         $this->assertEquals('SdcComponent', $metadata->getName());
@@ -61,7 +59,7 @@ class AsSdcComponentIntegrationTest extends KernelTestCase
         $this->assertEquals('js', $assets[1]['type']);
     }
 
-    public function testSdcComponentWithAssetAttribute(): void
+    public function testSdcComponentWithMultipleSdcAssetAttributes(): void
     {
         $kernel = new TestKernel([]);
         $kernel->boot();
@@ -73,11 +71,8 @@ class AsSdcComponentIntegrationTest extends KernelTestCase
 
         $this->assertCount(3, $assets);
 
-        // From #[Asset]
-        $this->assertEquals('css/extra.css', $assets[0]['path']);
-
-        // From #[AsSdcComponent]
-        $this->assertEquals('css/sdc.css', $assets[1]['path']);
-        $this->assertEquals('js/sdc.js', $assets[2]['path']);
+        $this->assertEquals('css/sdc.css', $assets[0]['path']);
+        $this->assertEquals('js/sdc.js', $assets[1]['path']);
+        $this->assertEquals('css/extra.css', $assets[2]['path']);
     }
 }