+++ tests/Integration/Fixtures/Component/AutoDiscovery/AutoDiscoveryComponent.php
@@ -12,10 +12,10 @@
 namespace Tito10047\UX\Sdc\Tests\Integration\Fixtures\Component\AutoDiscovery;
 
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
-use Tito10047\UX\Sdc\Attribute\Asset;
+use Tito10047\UX\Sdc\Attribute\AsSdcComponent;
 
 #[AsTwigComponent()]
-#[Asset]
+#[AsSdcComponent]
 class AutoDiscoveryComponent
 {
 }