+++ tests/Integration/Fixtures/Component/AutoDiscovery/SdcAutoDiscoveryComponent.php
@@ -11,9 +11,11 @@
 
 namespace Tito10047\UX\Sdc\Tests\Integration\Fixtures\Component\AutoDiscovery;
 
+use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 use Tito10047\UX\Sdc\Attribute\AsSdcComponent;
 
-#[AsSdcComponent('SdcAutoDiscoveryComponent')]
+#[AsTwigComponent('SdcAutoDiscoveryComponent')]
+#[AsSdcComponent]
 class SdcAutoDiscoveryComponent
 {
 }