+++ tests/Integration/Fixtures/Component/LiveComponentWithAsset.php
@@ -12,14 +12,13 @@
 namespace Tito10047\UX\Sdc\Tests\Integration\Fixtures\Component;
 
 use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
-use Symfony\UX\LiveComponent\DefaultActionTrait;
-use Tito10047\UX\Sdc\Attribute\Asset;
+use Tito10047\UX\Sdc\Attribute\AsSdcComponent;
 use Tito10047\UX\Sdc\Twig\ComponentNamespaceInterface;
 use Tito10047\UX\Sdc\Twig\Stimulus;
 
 if (trait_exists(\Symfony\UX\LiveComponent\DefaultActionTrait::class)) {
     #[AsLiveComponent('LiveComponentWithAsset')]
-    #[Asset]
+    #[AsSdcComponent]
     class LiveComponentWithAsset implements ComponentNamespaceInterface
     {
         use \Symfony\UX\LiveComponent\DefaultActionTrait;