+++ tests/Integration/Fixtures/Component/TestComponent.php
@@ -12,11 +12,11 @@
 namespace Tito10047\UX\Sdc\Tests\Integration\Fixtures\Component;
 
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
-use Tito10047\UX\Sdc\Attribute\Asset;
+use Tito10047\UX\Sdc\Attribute\SdcAsset;
 
 #[AsTwigComponent('TestComponent', template: 'components/TestComponent.html.twig')]
-#[Asset(path: 'css/test.css')]
-#[Asset(path: 'js/test.js')]
+#[SdcAsset(path: 'css/test.css')]
+#[SdcAsset(path: 'js/test.js')]
 class TestComponent
 {
 }