+++ tests/Unit/CompilerPass/AssetComponentCompilerPassTest.php
@@ -14,11 +14,11 @@ namespace Tito10047\UX\Sdc\Tests\Unit\CompilerPass;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
-use Tito10047\UX\Sdc\Attribute\Asset;
+use Tito10047\UX\Sdc\Attribute\AsSdcComponent;
 use Tito10047\UX\Sdc\CompilerPass\AssetComponentCompilerPass;
 use Tito10047\UX\Sdc\Runtime\SdcMetadataRegistry;
 
-#[Asset(path: 'test.css', priority: 5)]
+#[AsSdcComponent(path: 'test.css', priority: 5)]
 class MockComponent
 {
 }